@@ -16,11 +16,11 @@ module Integral
     # Presents blog tags
     def show
       add_breadcrumb t('integral.breadcrumbs.tags'), integral.tags_url
-      add_breadcrumb @tag.name, nil
+      add_breadcrumb @resource.name, nil
 
-      page_title = "Posts tagged with #{@tag.name}"
+      page_title = "Posts tagged with #{@resource.name}"
       page_title += " - Page #{params[:page]}" if params[:page].present?
-      page_description = params[:page].present? ? "Page #{params[:page] } - #{t('integral.tags.show.description', tag_name: @tag.name)}" : t('integral.tags.show.description', tag_name: @tag.name)
+      page_description = params[:page].present? ? "Page #{params[:page] } - #{t('integral.tags.show.description', tag_name: @resource.name)}" : t('integral.tags.show.description', tag_name: @resource.name)
 
       @meta_data = {
         page_title: page_title,
@@ -31,11 +31,11 @@ module Integral
     private
 
     def set_resource
-      @tag = Integral::Post.tag_counts_on("published_#{I18n.locale}").find_by_name!(params[:id])
+      @resource = Integral::Post.tag_counts_on("published_#{I18n.locale}").find_by_name!(params[:id])
     end
 
     def set_collection
-      @tagged_posts = PostDecorator.decorate_collection(Integral::Post.tagged_with(@tag.name).where(locale: I18n.locale).published.distinct.order("published_at DESC").paginate(:page => params[:page]))
+      @tagged_posts = PostDecorator.decorate_collection(Integral::Post.tagged_with(@resource.name).where(locale: I18n.locale).published.distinct.order("published_at DESC").paginate(:page => params[:page]))
     end
 
     def set_breadcrumbs
@@ -44,7 +44,7 @@ module Integral
     end
 
     def validate_page_has_results
-      raise_pagination_out_of_range if @tagged_posts.empty? && params[:page].present?
+      raise_pagination_out_of_range if @tagged_posts.empty?
     end
   end
 end
