@@ -16,7 +16,7 @@ module Integral
       private
 
       def white_listed_grid_params
-        %i[descending order page title]
+        %i[descending order page search]
       end
 
       def resource_klass
