@@ -1,6 +1,8 @@
 module Integral
   # Handles visitor communications
   class ContactController < Integral.frontend_parent_controller.constantize
+    before_action :enforce_honeypot, only: [:contact]
+
     # POST /contact
     def contact
       @enquiry = Enquiry.new(contact_params)
@@ -26,11 +28,15 @@ module Integral
     private
 
     def contact_params
-      params.require(:enquiry).permit(:name, :email, :message, :subject, :context, :newsletter_opt_in)
+      params.require(:enquiry).permit(:first_name, :name, :email, :message, :subject, :context, :newsletter_opt_in)
     end
 
     def newsletter_signup_params
       params.require(:newsletter_signup).permit(:email, :name, :context)
+    end
+
+    def enforce_honeypot
+      head(204) if contact_params[:first_name].present?
     end
   end
 end
