@@ -1,8 +1,7 @@
 module Integral
   # Posts controller
   class PostsController < BlogController
-    before_action :find_post, only: [:show]
-    before_action :find_related_posts, only: [:show]
+    before_action :set_resource, only: [:show]
     before_action :validate_page_has_results, only: [:index]
     after_action :increment_post_count, only: [:show]
 
@@ -32,13 +31,13 @@ module Integral
     # Presents blog postings
     def show
       add_breadcrumb I18n.t('integral.breadcrumbs.blog'), integral.posts_url
-      add_breadcrumb @post.title, nil
+      add_breadcrumb @resource.title, nil
 
       @meta_data = {
-        page_title: @post.title,
-        page_description: @post.description,
+        page_title: @resource.title,
+        page_description: @resource.description,
         open_graph: {
-          image: @post.preview_image_url(size: :large)
+          image: @resource.preview_image_url(size: :large)
         }
       }
       template = 'default' # TODO: Implement post templates
@@ -53,7 +52,7 @@ module Integral
       }
 
       if action_name == 'show'
-        @post.alternates.published.each do |alternate|
+        @resource.alternates.published.each do |alternate|
           alternative_urls[alternate.locale] = alternate.frontend_url
         end
       elsif
@@ -65,36 +64,40 @@ module Integral
     end
 
     def increment_post_count
-      @post.increment_count!(request.ip)
+      @resource.increment_count!(request.ip)
     end
 
     # Creates array of related posts. If enough related posts do not exist uses popular posts
-    def find_related_posts
-      @related_posts = @post.find_related_tags.limit(amount_of_related_posts_to_display)
-      amount_of_related_posts = @related_posts.length
+    def related_blog_posts
+      @related_posts ||= begin
+                           posts = @resource.find_related_tags.limit(amount_of_related_posts_to_display)
+                           amount_of_related_posts = posts.length
 
-      if amount_of_related_posts != amount_of_related_posts_to_display && (amount_of_related_posts + @popular_posts.length) >= amount_of_related_posts_to_display
-        @related_posts = @related_posts.to_a.concat(@popular_posts[0...amount_of_related_posts_to_display - amount_of_related_posts])
-      end
+                           if amount_of_related_posts != amount_of_related_posts_to_display && (amount_of_related_posts + popular_blog_posts.length) >= amount_of_related_posts_to_display
+                             posts = posts.to_a.concat(popular_blog_posts[0...amount_of_related_posts_to_display - amount_of_related_posts])
+                           end
+                           posts
+                         end
     end
+    helper_method :related_blog_posts
 
     def amount_of_related_posts_to_display
       3
     end
 
-    def find_post
-      @post = if current_user.present?
+    def set_resource
+      @resource = if current_user.present?
                 Integral::Post.where(locale: I18n.locale).friendly.find(params[:id]).decorate
               else
                 Integral::Post.where(locale: I18n.locale).friendly.published.find(params[:id]).decorate
               end
 
-      @post.decorate
+      @resource.decorate
 
       # If an old id or a numeric id was used to find the record, then
       # the request path will not match the post_path, and we should do
       # a 301 redirect that uses the current friendly id.
-      redirect_to integral.post_url(@post.slug), status: :moved_permanently if request.path != integral.post_path(@post.slug)
+      redirect_to integral.post_url(@resource.slug), status: :moved_permanently if request.path != integral.post_path(@resource.slug)
     end
 
     def validate_page_has_results
