@@ -18,8 +18,8 @@ module Integral
       object.posts.published.order('published_at DESC').limit(limit).decorate
     end
 
-    def image_url(size: nil, transform: nil)
-      image_variant_url(image, size: size, transform: transform)
+    def image_url(size: nil, transform: nil, fallback: true)
+      image_variant_url(image, size: size, transform: transform, fallback: fallback)
     end
   end
 end
