@@ -9,9 +9,8 @@ module Integral
     # @param enquiry [Enquiry] enquiry which the visitor has created
     def forward_enquiry(enquiry)
       @enquiry = enquiry
-      sender = email_sender(enquiry.name, enquiry.email)
 
-      mail subject: forwarding_subject(enquiry), from: sender, to: incoming_email_address, reply_to: sender
+      mail subject: forwarding_subject(enquiry), from: email_sender(enquiry.name, outgoing_email_address), to: incoming_email_address, reply_to: email_sender(enquiry.name, enquiry.email)
     end
 
     # Send an auto reply to the visitor of the supplied enquiry
@@ -19,6 +18,9 @@ module Integral
     # @param enquiry [Enquiry] enquiry which the visitor has created
     def auto_reply(enquiry)
       @enquiry = enquiry
+
+      I18n.locale = @enquiry.context
+
       sender = email_sender(Integral::Settings.website_title, outgoing_email_address)
 
       mail subject: auto_reply_subject(enquiry), to: enquiry.email, from: sender, reply_to: sender
