@@ -34,10 +34,16 @@ module Integral
     def render
       return render_no_object_warning if list_item.object? && !object_available?
 
-      content_tag opts[:wrapper_element], class: html_classes do
-        if list_item.has_children?
+      if list_item.has_children?
+        content_tag opts[:wrapper_element], class: html_classes do
           concat render_item
           concat content_tag opts[:child_wrapper_element], render_children, { class: opts[:child_wrapper_class] }, false
+        end
+      else
+        if opts[:wrapper_element]
+          content_tag opts[:wrapper_element], class: html_classes do
+            render_item
+          end
         else
           render_item
         end
@@ -117,7 +123,7 @@ module Integral
       image = object_data[:image] if object_available?
 
       if image.respond_to?(:attached?) && image.attached?
-        url_helpers.url_for(image.representation(resize_to_limit: Integral.image_sizes[size]))
+        url_helpers.url_for(image.representation(Integral.image_transformation_options.merge!(resize_to_limit: Integral.image_sizes[size])))
       elsif image.present?
         image
       else
@@ -132,7 +138,7 @@ module Integral
     # Returns the non object image path
     def non_object_image_url(size: :medium)
       if non_object_image.present?
-        url_helpers.url_for(non_object_image.representation(resize_to_limit: Integral.image_sizes[size]))
+        url_helpers.url_for(non_object_image.representation(Integral.image_transformation_options.merge!(resize_to_limit: Integral.image_sizes[size])))
       else
         fallback_image_url
       end
@@ -146,11 +152,15 @@ module Integral
       end
     end
 
-    def image_url(size: :medium)
+    def image_url(size: :medium, fallback: true)
       if image.present?
-        url_helpers.url_for(image.representation(resize_to_limit: Integral.image_sizes[size]))
+        if size == :original
+          url_helpers.url_for(image.attachment)
+        else
+          url_helpers.url_for(image.representation(Integral.image_transformation_options.merge!(resize_to_limit: Integral.image_sizes[size])))
+        end
       else
-        fallback_image_url
+        fallback_image_url if fallback
       end
     end
 
@@ -163,6 +173,16 @@ module Integral
     # @return [String] path to fallback image for list items
     def fallback_image_url
       ActionController::Base.helpers.image_path('integral/defaults/no_image_available.jpg')
+    end
+
+    def object_data
+      return nil unless object_available?
+
+      @object_data ||= list_item.object.to_list_item
+    end
+
+    def video?
+      object_data&.fetch(:image)&.video?
     end
 
     private
@@ -192,10 +212,6 @@ module Integral
 
       # Provide empty string - no attr supplied and no object linked
       ''
-    end
-
-    def object_data
-      @object_data ||= list_item.object.to_list_item
     end
 
     def object_available?
