@@ -5,10 +5,11 @@ module Integral
     routes { Integral::Engine.routes }
 
     let!(:resource) { create(:integral_category) }
+    let!(:post) { create(:integral_post, category: resource, status: 1) }
 
     describe 'GET show' do
       before do
-        get :show, params: { id: resource.to_param }
+        get :show, params: { id: resource.to_param, integral_original_path: category_path(resource.to_param) }
       end
 
       it { expect(response.status).to eq 200 }
