@@ -27,7 +27,7 @@ module Integral
 
     describe '.integral_icon' do
       it 'returns expected icon' do
-        expect(described_class.integral_icon).to eq 'user'
+        expect(described_class.integral_icon).to eq 'bi bi-people'
       end
     end
 
